+++ cli/src/dag/validator.rs
@@ -5,6 +5,8 @@ use {
     std::collections::{HashMap, HashSet},
 };
 
+type GraphAndPortEntryGroups = (DiGraph<GraphNode, ()>, HashMap<GraphNode, Vec<String>>);
+
 /// Validate function takes a DAG and validates it based on nexus execution
 /// rules.
 ///
@@ -13,7 +15,7 @@ use {
 /// <https://talus-labs.gitbook.io/talus-documentation/devs/index-1/cli#nexus-dag>
 pub(crate) fn validate(dag: Dag) -> AnyResult<()> {
     // Parse the dag into a petgraph DiGraph.
-    let graph: DiGraph<GraphNode, ()> = try_into_graph(dag)?;
+    let (graph, port_entry_groups) = try_into_graph(dag)?;
 
     if !graph.is_directed() || petgraph::algo::is_cyclic_directed(&graph) {
         bail!("The provided graph contains one or more cycles.");
@@ -23,7 +25,7 @@ pub(crate) fn validate(dag: Dag) -> AnyResult<()> {
     has_correct_order_of_actions(&graph)?;
 
     // Check that no walks in the graph violate the concurrency rules.
-    follows_concurrency_rules(&graph)?;
+    follows_concurrency_rules(&graph, &port_entry_groups)?;
 
     Ok(())
 }
@@ -74,41 +76,38 @@ fn has_correct_order_of_actions(graph: &DiGraph<GraphNode, ()>) -> AnyResult<()>
     Ok(())
 }
 
-/// For each distinct group of entry vertices, check that the net concurrency
+/// For each distinct group of entry input ports, check that the net concurrency
 /// leading from these nodes into any input port is always 0.
-fn follows_concurrency_rules(graph: &DiGraph<GraphNode, ()>) -> AnyResult<()> {
-    // Get all distinct groups of entry vertices.
-    let groups = graph
-        .node_indices()
-        .flat_map(|node| match &graph[node] {
-            GraphNode::Vertex { groups, .. } => groups.clone().unwrap_or_default(),
-            _ => vec![],
-        })
-        .collect::<HashSet<_>>();
+fn follows_concurrency_rules(
+    graph: &DiGraph<GraphNode, ()>,
+    port_entry_groups: &HashMap<GraphNode, Vec<String>>,
+) -> AnyResult<()> {
+    // Get all distinct groups of entry input ports.
+    let groups = port_entry_groups.values().flatten().collect::<HashSet<_>>();
 
     // For each group...
     for group in groups {
-        // ... find the entry vertices in that group.
-        let entry_vertices = graph
+        // ... find the entry input ports in that group.
+        let entry_input_ports = graph
             .node_indices()
-            .filter(|&node| match &graph[node] {
-                GraphNode::Vertex {
-                    groups: Some(groups),
-                    ..
-                } => groups.contains(&group),
-                _ => false,
+            .filter(|&node| {
+                port_entry_groups
+                    .get(&graph[node])
+                    .unwrap_or(&vec![])
+                    .contains(group)
             })
             .collect::<Vec<_>>();
 
         let input_ports = graph
             .node_indices()
-            .filter(|&node| matches!(graph[node], GraphNode::InputPort { .. }));
+            // Ignore entry input ports.
+            .filter(|&node| matches!(graph[node], GraphNode::InputPort { .. }) && !port_entry_groups.contains_key(&graph[node]));
 
         // And then for each input port ...
         for input_port in input_ports {
             // ... find all nodes that are included in the paths leading to
             // the input port.
-            let all_nodes_in_paths = entry_vertices
+            let all_nodes_in_paths = entry_input_ports
                 .iter()
                 .flat_map(|&entry_vertex| {
                     let min_intermediate_nodes = 0;
@@ -134,7 +133,15 @@ fn follows_concurrency_rules(graph: &DiGraph<GraphNode, ()>) -> AnyResult<()> {
                 );
             }
 
-            let concurrency = get_net_concurrency_in_subgraph(graph, &all_nodes_in_paths);
+            // Initial concurrency is the number of entry input ports in paths
+            // leading to the input port.
+            let initial_concurrency = entry_input_ports
+                .iter()
+                .filter(|&entry_vertex| all_nodes_in_paths.contains(entry_vertex))
+                .count() as isize;
+
+            let concurrency =
+                get_net_concurrency_in_subgraph(graph, &all_nodes_in_paths, initial_concurrency);
 
             if concurrency < 0 {
                 bail!(
@@ -160,8 +167,9 @@ fn follows_concurrency_rules(graph: &DiGraph<GraphNode, ()>) -> AnyResult<()> {
 fn get_net_concurrency_in_subgraph(
     graph: &DiGraph<GraphNode, ()>,
     nodes: &HashSet<NodeIndex>,
+    initial_concurrency: isize,
 ) -> isize {
-    let net_concurrency = nodes.iter().fold(0, |acc, &node| {
+    let net_concurrency = nodes.iter().fold(initial_concurrency, |acc, &node| {
         match graph[node] {
             GraphNode::Vertex { .. } => {
                 // Calculate the maximum number of concurrent tasks that can be spawned by this tool.
@@ -184,7 +192,7 @@ fn get_net_concurrency_in_subgraph(
                 // Add 1 as we only want to consume concurrency if there's more than 1 input port.
                 acc + max_tool_concurrency + 1
             }
-            // Input ports with no default values reduce concurrency.
+            // Input ports reduce concurrency.
             GraphNode::InputPort { .. } => acc - 1,
             _ => acc,
         }
@@ -202,12 +210,6 @@ enum GraphNode {
     },
     Vertex {
         name: String,
-        /// `groups` is:
-        /// - `Some(vec) if !vec.is_empty()` when it's an entry vertex
-        /// - `Some(vec) if vec.is_empty()` is unreachable as default group is
-        ///   provided in case one is missing
-        /// - `None` when it's a normal vertex
-        groups: Option<Vec<String>>,
     },
     OutputVariant {
         vertex: String,
@@ -224,7 +226,7 @@ impl std::fmt::Display for GraphNode {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             GraphNode::InputPort { vertex, name } => write!(f, "Input port: {vertex}.{name}"),
-            GraphNode::Vertex { name, .. } => write!(f, "Vertex: {name}"),
+            GraphNode::Vertex { name } => write!(f, "Vertex: {name}"),
             GraphNode::OutputVariant { vertex, name } => {
                 write!(f, "Output variant: {vertex}.{name}")
             }
@@ -241,73 +243,75 @@ impl std::fmt::Display for GraphNode {
 
 /// [Dag] to [petgraph::graph::DiGraph]. Also performs structure checks on the
 /// graph.
-fn try_into_graph(dag: Dag) -> AnyResult<DiGraph<GraphNode, ()>> {
+fn try_into_graph(dag: Dag) -> AnyResult<GraphAndPortEntryGroups> {
     let mut graph = DiGraph::<GraphNode, ()>::new();
 
-    // Check that there is at least one entry vertex.
-    if dag.entry_vertices.is_empty() {
-        bail!("The DAG has no entry vertices.");
-    }
+    // Build a hash map of graph nodes that are part of entry groups. If there
+    // are no entry groups specified, we assume all specified input ports are
+    // part of the default entry group.
+    let mut port_entry_groups: HashMap<GraphNode, Vec<String>> = HashMap::new();
 
-    // For each entry vertex, find a set of entry groups it belongs to.
-    let mut vertex_entry_groups: HashMap<String, Vec<String>> = HashMap::new();
+    if let Some(entry_groups) = &dag.entry_groups {
+        for entry_group in entry_groups {
+            for member in &entry_group.members {
+                let port = GraphNode::InputPort {
+                    vertex: member.vertex.clone(),
+                    // Entry vertices with no input ports will get a phantom
+                    // input port.
+                    name: member.input_port.clone().unwrap_or("__phantom".to_string()),
+                };
 
-    for entry_vertex in &dag.entry_vertices {
-        match &dag.entry_groups {
-            None => {
-                vertex_entry_groups.insert(
-                    entry_vertex.name.clone(),
-                    vec![DEFAULT_ENTRY_GROUP.to_string()],
-                );
-            }
-            Some(entry_groups) => {
-                let entry_groups = entry_groups
-                    .iter()
-                    .filter(|entry_group| entry_group.vertices.contains(&entry_vertex.name))
-                    .map(|entry_group| entry_group.name.clone())
-                    .collect::<Vec<String>>();
+                // Check that if alone vertices are specified, they don't have
+                // input ports.
+                if member.input_port.is_none() {
+                    let Some(vertex) = dag
+                        .vertices
+                        .iter()
+                        .find(|vertex| vertex.name == member.vertex)
+                    else {
+                        bail!(
+                            "Entry group '{}' references a non-existing vertex '{}'.",
+                            entry_group.name,
+                            member.vertex
+                        );
+                    };
 
-                // If the entry vertex does not belong to any group, add it
-                // to the default group.
-                match entry_groups.len() {
-                    0 => {
-                        vertex_entry_groups.insert(
-                            entry_vertex.name.clone(),
-                            vec![DEFAULT_ENTRY_GROUP.to_string()],
+                    if vertex.input_ports.is_some() {
+                        bail!(
+                            "Entry group '{}' references a vertex '{}' that has input ports.",
+                            entry_group.name,
+                            member.vertex
                         );
                     }
-                    _ => {
-                        vertex_entry_groups.insert(entry_vertex.name.clone(), entry_groups);
-                    }
                 }
+
+                let mut groups = port_entry_groups.remove(&port).unwrap_or_default();
+                groups.push(entry_group.name.clone());
+                port_entry_groups.insert(port, groups);
+            }
+        }
+    } else {
+        // If there are no entry groups, all input ports are part of the
+        // default entry group.
+        for vertex in &dag.vertices {
+            let Some(input_ports) = vertex.input_ports.as_ref() else {
+                continue;
+            };
+
+            for input_port in input_ports {
+                let port = GraphNode::InputPort {
+                    vertex: vertex.name.clone(),
+                    name: input_port.clone(),
+                };
+
+                port_entry_groups.insert(port, vec![DEFAULT_ENTRY_GROUP.to_string()]);
             }
         }
     }
 
-    // Check that entry groups only reference entry vertices and that entry
-    // group names don't use the reserved keyword.
-    let entry_groups = dag.entry_groups.unwrap_or_default();
-
-    for entry_group in &entry_groups {
-        if entry_group.name == DEFAULT_ENTRY_GROUP {
-            bail!(
-                "Entry group '{}' uses a reserved keyword.",
-                DEFAULT_ENTRY_GROUP
-            );
-        }
-
-        for vertex in &entry_group.vertices {
-            if !dag
-                .entry_vertices
-                .iter()
-                .any(|entry_vertex| entry_vertex.name == *vertex)
-            {
-                bail!(
-                    "Entry group '{}' references a non-entry 'Vertex: {vertex}'.",
-                    entry_group.name
-                );
-            }
-        }
+    // Check that there is at least one entry point.
+    if port_entry_groups.is_empty() {
+        bail!("The DAG has no entry vertices or ports.");
     }
 
     // Edges are always between an output port and an input port. We also
@@ -318,7 +322,6 @@ fn try_into_graph(dag: Dag) -> AnyResult<DiGraph<GraphNode, ()>> {
     for edge in dag.edges {
         let origin_vertex = GraphNode::Vertex {
             name: edge.from.vertex.clone(),
-            groups: vertex_entry_groups.get(&edge.from.vertex).cloned(),
         };
 
         let output_variant = GraphNode::OutputVariant {
@@ -334,7 +337,6 @@ fn try_into_graph(dag: Dag) -> AnyResult<DiGraph<GraphNode, ()>> {
 
         let destination_vertex = GraphNode::Vertex {
             name: edge.to.vertex.clone(),
-            groups: vertex_entry_groups.get(&edge.to.vertex).cloned(),
         };
 
         let input_port = GraphNode::InputPort {
@@ -412,46 +414,37 @@ fn try_into_graph(dag: Dag) -> AnyResult<DiGraph<GraphNode, ()>> {
     }
 
     // Ensure we don't have duplicate vertices.
-    let mut all_entry_vertices = HashSet::new();
     let mut all_vertices = HashSet::new();
-    let mut all_entry_input_ports = HashSet::new();
 
-    // Check that all entry vertices are in the graph. Note that connecting
-    // entry input ports to these entry vertices is not necessary as they do
-    // not matter for the validation.
-    for entry_vertex in &dag.entry_vertices {
-        let entry_vertex_ident = GraphNode::Vertex {
-            name: entry_vertex.name.clone(),
-            groups: vertex_entry_groups.get(&entry_vertex.name).cloned(),
+    // Connect entry input ports and vertices to the graph for validation. Note
+    // that these ports are guaranteed to not exist on the graph yet.
+    for input_port in port_entry_groups.keys() {
+        let GraphNode::InputPort { vertex, name } = input_port else {
+            unreachable!();
         };
 
-        if !graph_nodes.contains_key(&entry_vertex_ident) {
+        let entry_vertex_ident = GraphNode::Vertex {
+            name: vertex.clone(),
+        };
+
+        let entry_input_port_ident = GraphNode::InputPort {
+            vertex: vertex.clone(),
+            name: name.clone(),
+        };
+
+        let Some(vertex_node) = graph_nodes.get(&entry_vertex_ident) else {
             bail!("Entry '{entry_vertex_ident}' is not connected to the DAG.",);
-        }
+        };
 
-        if !all_entry_vertices.insert(entry_vertex_ident.clone()) {
-            bail!("Entry '{entry_vertex_ident}' is defined multiple times.",);
-        }
+        let input_port_node = graph.add_node(entry_input_port_ident);
 
-        // Add entry input ports to the map so we can check that they do not
-        // have a default value.
-        for input_port in &entry_vertex.input_ports {
-            let input_port_ident = GraphNode::InputPort {
-                vertex: entry_vertex.name.clone(),
-                name: input_port.clone(),
-            };
-
-            if !all_entry_input_ports.insert(input_port_ident.clone()) {
-                bail!("Entry '{input_port_ident}' is defined multiple times.",);
-            }
-        }
+        graph.add_edge(input_port_node, *vertex_node, ());
     }
 
     // Check that all normal vertices are in the graph.
     for vertex in &dag.vertices {
         let vertex_ident = GraphNode::Vertex {
             name: vertex.name.clone(),
-            groups: vertex_entry_groups.get(&vertex.name).cloned(),
         };
 
         if !graph_nodes.contains_key(&vertex_ident) {
@@ -461,11 +454,6 @@ fn try_into_graph(dag: Dag) -> AnyResult<DiGraph<GraphNode, ()>> {
         if !all_vertices.insert(vertex_ident.clone()) {
             bail!("'{vertex_ident}' is defined multiple times.",);
         }
-    }
-
-    // Ensure vertex is not specified as a vertex and an entry vertex.
-    if let Some(vertex) = all_vertices.intersection(&all_entry_vertices).next() {
-        bail!("'{vertex}' is both a vertex and an entry vertex.")
     }
 
     // Check that none of the default value input ports are in the graph.
@@ -478,13 +466,13 @@ fn try_into_graph(dag: Dag) -> AnyResult<DiGraph<GraphNode, ()>> {
         };
 
         if graph_nodes.contains_key(&default_value)
-            || all_entry_input_ports.contains(&default_value)
+            || port_entry_groups.contains_key(&default_value)
         {
             bail!(
-                    "'{default_value}' is already present in the graph or has an edge leading into it and therefore cannot have a default value.",
-                );
+                "'{default_value}' is already present in the graph or has an edge leading into it and therefore cannot have a default value.",
+            );
         }
     }
 
-    Ok(graph)
+    Ok((graph, port_entry_groups))
 }