+++ cli/src/tool/tool_register.rs
@@ -3,11 +3,12 @@ use {
         command_title,
         display::json_output,
         loading,
+        notify_success,
         prelude::*,
         sui::*,
         tool::{tool_validate::*, ToolIdent},
     },
-    nexus_sdk::transactions::tool,
+    nexus_sdk::{idents::primitives, transactions::tool},
 };
 
 /// Validate and then register a new Tool.
@@ -33,6 +34,7 @@ pub(crate) async fn register_tool(
     // Nexus objects must be present in the configuration.
     let NexusObjects {
         workflow_pkg_id,
+        primitives_pkg_id,
         tool_registry_object_id,
         ..
     } = get_nexus_objects(&conf)?;
@@ -76,7 +78,7 @@ pub(crate) async fn register_tool(
 
     let mut tx = sui::ProgrammableTransactionBuilder::new();
 
-    match tool::register_off_chain(
+    match tool::register_off_chain_for_self(
         &mut tx,
         &meta,
         collateral_coin,
@@ -104,7 +106,38 @@ pub(crate) async fn register_tool(
     // Sign and submit the TX.
     let response = sign_transaction(&sui, &wallet, tx_data).await?;
 
-    json_output(&json!({ "digest": response.digest }))?;
+    // Parse the owner cap object ID from the response.
+    let owner_cap = response
+        .object_changes
+        .unwrap_or_default()
+        .into_iter()
+        .find_map(|change| match change {
+            sui::ObjectChange::Created {
+                object_type,
+                object_id,
+                ..
+            } if object_type.address == *primitives_pkg_id
+                && object_type.module
+                    == primitives::OwnerCap::CLONEABLE_OWNER_CAP.module.into()
+                && object_type.name == primitives::OwnerCap::CLONEABLE_OWNER_CAP.name.into() =>
+            {
+                Some(object_id)
+            }
+            _ => None,
+        });
+
+    let Some(object_id) = owner_cap else {
+        return Err(NexusCliError::Any(anyhow!(
+            "Could not find the OnwerCap object ID in the transaction response."
+        )));
+    };
+
+    notify_success!(
+        "OwnerCap object ID: {id}",
+        id = object_id.to_string().truecolor(100, 100, 100)
+    );
+
+    json_output(&json!({ "digest": response.digest, "owner_cap_id": object_id }))?;
 
     Ok(())
 }