+++ CHANGELOG.md
@@ -14,7 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - commands to validate, publish, execute and inspect DAGs
 - commands to load and save configuration
 - commands to create a new Nexus network
+
+### Changed
+
 - upgraded Sui from `testnet-1.38.1` to `mainnet-v1.45.3`
+- changing the notion of entry vertices to entry input ports and adjusting parsing, validation and PTB templates in accordance
+
+### Fixed
+
+- fixing tool registration, unregistration and collateral claiming based on changes in tool registry
 
 ## [`nexus-toolkit-rust` 0.1.0] - Unreleased
 
@@ -23,6 +31,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added basic strcuture for Nexus Tools written in Rust in the form of a trait
 - added a macro that starts a webserver for one or multiple tools, providing all necessary endpoints
 - added a first, dumb version of secret manager
+
+### Changed
+
 - upgraded Sui from `testnet-1.38.1` to `mainnet-v1.45.3`
 
 ## [`nexus-sdk` 0.1.0] - Unreleased
@@ -35,6 +46,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added `types` module and `tool_fqn` that holds some reusable types
 - added `events` module that holds definitions of Nexus events fired from Sui
 - added `sui` module that holds and categorizes all `sui_sdk` types
+
+### Changed
+
 - upgraded Sui from `testnet-1.38.1` to `mainnet-v1.45.3`
 
 ### Fixed